+++ packages/node/CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Fix modulo block didn't apply correctly with multiple dataSources (#2331)
+- Now when `workers` set to 0, it will use block dispatcher instead of throw and exit 
 
 ## [4.0.0] - 2024-03-28